requirements.txt
Flask==2.3.3
Flask-SocketIO==5.3.6
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
google-api-python-client==2.111.0
google-generativeai==0.3.1
python-dotenv==1.0.0
youtube-transcript-api==0.6.1
gunicorn==21.2.0
eventlet==0.33.3
python-engineio==4.8.0
python-socketio==5.10.0
Flask-Caching==2.1.0
Flask-WTF==1.2.1
email-validator==2.1.0.post1
